+++ rollup.config.js
@@ -4,6 +4,7 @@ import resolve from '@rollup/plugin-node-resolve';
 import livereload from 'rollup-plugin-livereload';
 import { terser } from 'rollup-plugin-terser';
 import css from 'rollup-plugin-css-only';
+import { uglify } from "rollup-plugin-uglify";
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -68,7 +69,9 @@ export default {
 
 		// If we're building for production (npm run build
 		// instead of npm run dev), minify
-		production && terser()
+		production && terser(),
+
+		uglify(),
 	],
 	watch: {
 		clearScreen: false